Split uploadData to public and private
