+++ background.html
@@ -138,7 +138,7 @@
                 </div>
                 <div class='choose-file'>
                     <button class='paperclip thumbnail' {{#disable-inputs}} disabled="disabled" {{/disable-inputs}}></button>
-                    <input type='file' class='file-input' multiple='multiple'>
+                    <input type='file' class='file-input' multiple='multiple' accept='video/* image/*'>
                 </div>
               </div>
             </form>
